// Intranet-Front/src/app/models/interface-email-dto.model.ts
export interface EmailJovemDTO {
  nomeJovem: string;        // Nome do aprendiz
  telefoneUsuario: string;   // Telefone do aprendiz ou responsável
  nomeUsuario: string;       // Nome do remetente
  nomeOrientador: string;    // Nome do responsável do RH da empresa
  emailDestino: string;      // Email para enviar
}